Merged in the multiple instance plone branch. Merge from: http://codespeak.net/svn/kupu/trunk/kupu, revision 9684 to http://codespeak.net/svn/kupu/branch/duncan-multiplone 9722.
Fixed a problem where multiple instances cause a runaway fight over the focus by removing the call to grab focus in KupuEditor.initialize: this may have an effect on other ports which expect Kupu to grab the focus, but if there is more than just Kupu on the page (whether its another Kupu or just a text area, then Kupu shouldn't assume it has the right to the focus).

Found an issue with zoom mode setting a negative height: worked around it, but some of the assumptions must be wrong somewhere.