Many changes to Plone installation.
If the resource registry is available, all of Kupu's css and js files are now loaded through the registry. If it is not available the old loading in the middle of the page still happens.

Removed browserSupportsKupu.py, and instead added a method isKupuEnabled to the tool.

Changed the name of the tool in the configlet as per Alex Limi's suggestion.