Plone ticket #6970 (can't upgrade to Kupu 1.4.1)
Added a workaround for a problem with PortalTransforms.

When reinstalling kupu the old transforms are removed and then added again. The identity transform captioned-to-html has parameters which need to be 
set after adding the transform, and it appears that if all three operations (remove, add and update) happen in the same transaction this error 
occurs.

I've added a savepoint to the transaction for the case where it creates the identity transform and that seems to fix the problem.